requirements.txt
python-dotenv
openai
spotipy